Pabaigta programėlė

+++ CO2Bakalauras/CO2Bakalauras/Models/Administratorius.cs
@@ -1,4 +1,6 @@
 ﻿
+using System;
+
 namespace CO2Bakalauras.Models
 {
     public class Administratorius
@@ -8,7 +10,7 @@ namespace CO2Bakalauras.Models
         public int VARTOTOJO_ID { get; set; }
         public int TELEFONO_NUMERIS { get; set; }
         public string ADRESAS { get; set; }
-        public int ASMENS_KODAS { get; set; }
+        public Int64 ASMENS_KODAS { get; set; }
         public bool BUSENA { get; set; }
     }
 }
