+++ CO2BakalaurasAPI/Models/AdministratoriusRequest.cs
@@ -6,7 +6,7 @@
         public int VARTOTOJO_ID { get; set; }
         public int TELEFONO_NUMERIS { get; set; }
         public string? ADRESAS { get; set; }
-        public int ASMENS_KODAS { get; set; }
+        public Int64 ASMENS_KODAS { get; set; }
         public bool BUSENA { get; set; }
 
     }
